extracted data types to separate file
